app/models/db.py
"""
数据库模型
(预留用于未来数据库集成)
"""

from typing import Optional
from datetime import datetime

# 注意: 这里使用伪代码，实际使用时需要安装 SQLAlchemy 或其他 ORM

# from sqlalchemy import Column, Integer, String, DateTime, Text, Boolean
# from sqlalchemy.ext.declarative import declarative_base
# 
# Base = declarative_base()
# 
# class Document(Base):
#     """文档表"""
#     __tablename__ = "documents"
#     
#     id = Column(Integer, primary_key=True, index=True)
#     document_id = Column(String(50), unique=True, index=True)
#     filename = Column(String(255))
#     file_path = Column(String(500))
#     file_type = Column(String(10))  # pdf, docx
#     file_size = Column(Integer)
#     upload_time = Column(DateTime, default=datetime.now)
#     status = Column(String(20), default="uploaded")  # uploaded, processing, completed, failed
#     
#     # 元数据
#     title = Column(String(500))
#     author = Column(String(255))
#     page_count = Column(Integer)
#     text_length = Column(Integer)
#     
#     # 分析结果
#     structure_analyzed = Column(Boolean, default=False)
#     images_extracted = Column(Boolean, default=False)
#     ppt_generated = Column(Boolean, default=False)
# 
# class Image(Base):
#     """图像表"""
#     __tablename__ = "images"
#     
#     id = Column(Integer, primary_key=True, index=True)
#     image_id = Column(String(50), unique=True, index=True)
#     document_id = Column(String(50), index=True)
#     name = Column(String(255))
#     file_path = Column(String(500))
#     page_number = Column(Integer)
#     image_type = Column(String(20))  # chart, diagram, photo, equation
#     format = Column(String(10))  # jpg, png, etc.
#     size = Column(Integer)
#     caption = Column(Text)
#     created_at = Column(DateTime, default=datetime.now)
# 
# class Analysis(Base):
#     """分析结果表"""
#     __tablename__ = "analyses"
#     
#     id = Column(Integer, primary_key=True, index=True)
#     document_id = Column(String(50), index=True)
#     analysis_type = Column(String(50))  # structure, keyword, summary
#     result = Column(Text)  # JSON 格式的结果
#     created_at = Column(DateTime, default=datetime.now)
# 
# class Presentation(Base):
#     """PPT 记录表"""
#     __tablename__ = "presentations"
#     
#     id = Column(Integer, primary_key=True, index=True)
#     ppt_id = Column(String(50), unique=True, index=True)
#     document_id = Column(String(50), index=True)
#     file_path = Column(String(500))
#     template = Column(String(50))
#     slide_count = Column(Integer)
#     created_at = Column(DateTime, default=datetime.now)
#     download_count = Column(Integer, default=0)

# 数据库连接配置 (示例)
"""
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker

# 数据库连接字符串
DATABASE_URL = "sqlite:///./academic_paper_assistant.db"
# 或者使用 PostgreSQL:
# DATABASE_URL = "postgresql://user:password@localhost/dbname"

engine = create_engine(DATABASE_URL, echo=True)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()

# 创建所有表
Base.metadata.create_all(bind=engine)
"""

# 临时注释: 当前版本不使用数据库
# 未来版本可以取消注释并安装相关依赖

class DatabasePlaceholder:
    """数据库占位符类"""
    
    @staticmethod
    def get_info():
        return {
            "status": "not_configured",
            "message": "数据库功能暂未启用，当前版本使用内存存储",
            "future_features": [
                "持久化存储文档信息",
                "用户管理与权限控制",
                "分析历史记录",
                "高级搜索与过滤"
            ]
        }

# 导出占位符
db_info = DatabasePlaceholder()
