requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# 文档解析
PyMuPDF==1.23.8
python-docx==1.1.0

# PPT 生成
python-pptx==0.6.23

# 图像处理
Pillow==10.2.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# 可选依赖 (未来扩展)
# sqlalchemy==2.0.25
# alembic==1.13.1
# redis==5.0.1
# celery==5.3.4
# pandas==2.1.4
# numpy==1.26.3
